pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finance-sim"
version = "1.0.0"
description = "Personal Cash Flow Simulator & Reporter"
authors = [
    {name = "CD", email = "cd@cd"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas",
    "pytest>=8.4.0",
    "PyYAML",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-mock",
]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
